@@ -7,7 +7,7 @@
   "scripts": {
     "clean": "rm -rf ./build/*",
     "build": "npm run clean && tsc",
-    "lint": "tslint 'src/**/*.ts'",
+    "lint": "eslint 'src/**/*.ts'",
     "test": "jest",
     "test:coverage": "jest --coverage --collectCoverageFrom=src/**/*.ts",
     "test:watch": "jest --watchAll",
@@ -20,30 +20,31 @@
     "build"
   ],
   "dependencies": {
-    "@icapps/tree-house-errors": "~2.0.1",
-    "aws-sdk": "~2.605.0",
-    "express-validation": "~1.0.2",
-    "multer": "~1.4.2"
+    "@icapps/tree-house-errors": "~3.2.0",
+    "aws-sdk": "~2.1181.0",
+    "express-validation": "~4.1.0",
+    "multer": "~1.4.5-lts.1"
   },
   "devDependencies": {
-    "@types/cors": "~2.8.6",
-    "@types/express-brute": "~0.0.37",
-    "@types/helmet": "~0.0.45",
-    "@types/jest": "~24.9.0",
-    "@types/joi": "~14.3.4",
-    "@types/multer": "~1.3.10",
-    "@types/redis": "~2.8.14",
-    "coveralls": "~3.0.9",
-    "http-status": "~1.4.2",
-    "jest": "~24.9.0",
-    "node-mocks-http": "~1.8.1",
-    "np": "~5.2.1",
+    "@types/cors": "~2.8.12",
+    "@types/express": "~4.17.13",
+    "@types/express-brute": "~1.0.1",
+    "@types/helmet": "~4.0.0",
+    "@types/jest": "~28.1.6",
+    "@types/multer": "~1.4.7",
+    "@typescript-eslint/eslint-plugin": "~5.30.7",
+    "@typescript-eslint/parser": "~5.30.7",
+    "coveralls": "~3.1.1",
+    "eslint": "~8.20.0",
+    "eslint-config-prettier": "~8.5.0",
+    "http-status": "~1.5.2",
+    "jest": "~28.1.3",
+    "node-mocks-http": "~1.11.0",
+    "np": "~7.6.2",
     "pre-commit": "~1.2.2",
-    "supertest": "~4.0.2",
-    "ts-jest": "~24.3.0",
-    "tslint": "~5.20.1",
-    "tslint-config-airbnb": "~5.11.2",
-    "typescript": "~3.7.5"
+    "supertest": "~6.2.4",
+    "ts-jest": "~28.0.7",
+    "typescript": "~4.7.4"
   },
   "engines": {
     "node": ">=8.0.0"
